Moved check user exists to helpers
